Function re_smart_channel::smart_channel_with_stats

pub(crate) fn smart_channel_with_stats<T: Send>(
    sender_source: SmartMessageSource,
    source: Arc<SmartChannelSource>,
    stats: Arc<SharedStats>,
) -> (Sender<T>, Receiver<T>)

Create a new channel using the same stats as some other.

This is a very leaky abstraction, and it would be nice to refactor some day