Struct re_smart_channel::sender::Sender

pub struct Sender<T: Send> {
    tx: Sender<SmartMessage<T>>,
    source: Arc<SmartMessageSource>,
    stats: Arc<SharedStats>,
}

Fields
tx: Sender<SmartMessage<T>>source: Arc<SmartMessageSource>stats: Arc<SharedStats>
Implementations

impl<T: Send> Sender<T>

pub(crate) fn new( tx: Sender<SmartMessage<T>>, source: Arc<SmartMessageSource>, stats: Arc<SharedStats> ) -> Self

pub fn clone_as(&self, source: SmartMessageSource) -> Self

Clones the sender with an updated source.

pub fn send(&self, msg: T) -> Result<(), SendError<T>>

pub fn send_at( &self, time: Instant, source: Arc<SmartMessageSource>, payload: SmartMessagePayload<T> ) -> Result<(), SendError<SmartMessagePayload<T>>>

Forwards a message as-is.

pub fn flush_blocking(&self) -> Result<(), SendError<()>>

Blocks until all previously sent messages have been received.

Note: This is only implemented for non-wasm targets since we cannot make blocking calls on web.

pub fn quit( &self, err: Option<Box<dyn Error + Send>> ) -> Result<(), SendError<SmartMessage<T>>>

Used to indicate that a sender has left.

This sends a message down the channel allowing the receiving end to know whether one of the sender has left, and if so why (if applicable).

Using a Sender after calling quit is undefined behavior: the receiving end is free to silently drop those messages (or worse).

pub fn is_empty(&self) -> bool

Is the channel currently empty of messages?

pub fn len(&self) -> usize

Number of messages in the channel right now.

pub fn latency_ns(&self) -> u64

Latest known latency from sending a message to receiving it, it nanoseconds.

pub fn latency_sec(&self) -> f32

Latest known latency from sending a message to receiving it, in seconds