Struct re_smart_channel::Receiver

pub struct Receiver<T: Send> {
    pub(crate) rx: Receiver<(Instant, T)>,
    pub(crate) stats: Arc<SharedStats>,
    pub(crate) source: Source,
}

Fields
rx: Receiver<(Instant, T)>stats: Arc<SharedStats>source: Source
Implementations

impl<T: Send> Receiver<T>

pub fn recv(&self) -> Result<T, RecvError>

pub fn try_recv(&self) -> Result<T, TryRecvError>

pub fn recv_timeout(&self, timeout: Duration) -> Result<T, RecvTimeoutError>

pub fn recv_with_send_time(&self) -> Result<(Instant, T), RecvError>

Receives without registering the latency.

This is for use with Sender::send_at when chaining to another channel created with Self::chained_channel.

pub fn source(&self) -> &Source

Where is the data coming from?

pub fn is_empty(&self) -> bool

Is the channel currently empty of messages?

pub fn len(&self) -> usize

Number of messages in the channel right now.

pub fn latency_ns(&self) -> u64

Latest known latency from sending a message to receiving it, it nanoseconds.

pub fn latency_sec(&self) -> f32

Latest known latency from sending a message to receiving it, in seconds

pub fn chained_channel(&self) -> (Sender<T>, Receiver<T>)

Create a new channel that use the same stats as this one.

This means both channels will see the same latency numbers.

Care must be taken to use Self::recv_with_send_time and Sender::send_at. This is a very leaky abstraction, and it would be nice with a refactor.