Crate re_smart_channel

A channel that keeps track of latency and queue length.

Modules

receive_set 🔒

receiver 🔒

sender 🔒

Structs

ReceiveSet

A set of connected Receivers.

Receiver

RecvError

An error returned from the recv method.

SendError

An error returned from the send method.

Sender

SharedStats 🔒

Stats for a channel, possibly shared between chained channels.

SmartMessage

Enums

RecvTimeoutError

An error returned from the recv_timeout method.

SmartChannelSource

Identifies in what context this smart channel was created, and who/what is holding its receiving end.

SmartMessagePayload

The payload of a SmartMessage.

SmartMessageSource

Identifies who/what sent a particular message in a smart channel.

TryRecvError

An error returned from the try_recv method.

Functions

smart_channel

smart_channel_with_stats 🔒

Create a new channel using the same stats as some other.